Allow Razor directives inside HTML attribute values #4
The HTML syntax has many contexts around single and double quoted attribute values.  We need to include a custom variant of the Razor keyword injector so it can switch back to HTML once it hits the end of the string.  We also need to clear the string scope to avoid any color scheme issues.

Additionally, keyword.declaration was renamed to keyword.other to better fit with standards.